// prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  PATIENT: 'PATIENT',
  ADMIN: 'ADMIN',
  DOCTOR: 'DOCTOR',
  CLINIC: 'CLINIC'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const AppointmentStatus = {
  REJECTED: 'REJECTED',
  CONFIRMED: 'CONFIRMED',
  PENDING: 'PENDING'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER',
  preferNotToSay: 'preferNotToSay'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const PaymentMethod = {
  online: 'online',
  cash: 'cash',
  other: 'other'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PaymentStatus = {
  CREATED: 'CREATED',
  PAID: 'PAID',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]